component arquitecture structure
draft, move all logic to their own components
